WFCORE-5899 Renamed the package name and name of the .java classes
