+++ testsuite/domain/pom.xml
@@ -197,7 +197,7 @@
                         <include>org/jboss/as/test/integration/domain/events/*TestCase.java</include>
                         <include>org/jboss/as/test/integration/domain/suites/*TestSuite.java</include>
                         <include>org/jboss/as/test/integration/domain/suspendresume/*TestCase.java</include>
-                        <include>org/jboss/as/test/integration/domain/slavereconnect/*TestCase.java</include>
+                        <include>org/jboss/as/test/integration/domain/secondaryreconnect/*TestCase.java</include>
                         <include>org/jboss/as/test/integration/domain/management/*TestCase.java</include>
                         <include>org/jboss/as/test/integration/domain/elytron/*TestCase.java</include>
                     </includes>
