+++ testsuite/domain/src/test/java/org/jboss/as/test/integration/domain/secondaryreconnect/OrderedChildResourceScenario.java
@@ -20,7 +20,7 @@
  * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
  */
 
-package org.jboss.as.test.integration.domain.slavereconnect;
+package org.jboss.as.test.integration.domain.secondaryreconnect;
 
 import static org.jboss.as.controller.descriptions.ModelDescriptionConstants.ADD_INDEX;
 import static org.jboss.as.controller.descriptions.ModelDescriptionConstants.EXTENSION;
